+++ src/hotspot/share/utilities/decoder_elf.hpp
@@ -43,7 +43,7 @@ public:
     ShouldNotReachHere();
     return false;
   }
-  bool get_source_info(address pc, char* buf, int buflen, bool is_pc_after_call);
+  bool get_source_info(address pc, char* buf, size_t buflen, bool is_pc_after_call);
 
 private:
   ElfFile*         get_elf_file(const char* filepath);
